requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
dj-database-url==2.3.0
Django==5.1.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.1
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.12.2
urllib3==2.3.0
gunicorn==23.0.0
flake8==7.3.0
